refactor: data-title-utils.js
